+++ Sudokus.ApiService/DTOs/Request.cs
@@ -1,5 +1,5 @@
 ﻿namespace Sudokus.ApiService.DTOs
 {
-    public record ComprobacionRequest (string IdSesion, int[,] TableroActual);
-    public record PistaCasillaRequest (string IdSesion, int fil, int col);
+    public record ComprobacionRequest (string IdSesion, byte[,] TableroActual);
+    public record PistaCasillaRequest (string IdSesion, byte fil, byte col);
 }
